[data] Refractor OpRuntimeMetrics
