+++ python/ray/data/_internal/execution/interfaces/physical_operator.py
@@ -25,7 +25,11 @@ from ray.data._internal.execution.interfaces.execution_options import (
     ExecutionOptions,
     ExecutionResources,
 )
-from ray.data._internal.execution.interfaces.op_runtime_metrics import OpRuntimeMetrics
+from ray.data._internal.execution.interfaces.op_runtime_metrics import (
+    BaseOpMetrics,
+)
+
+# Generic type variable for metrics
 from ray.data._internal.logical.interfaces import LogicalOperator, Operator
 from ray.data._internal.output_buffer import OutputBlockSizeOption
 from ray.data._internal.progress_bar import ProgressBar
@@ -339,8 +343,6 @@ class PhysicalOperator(Operator):
         )
         self._started = False
         self._shutdown = False
-        self._in_task_submission_backpressure = False
-        self._in_task_output_backpressure = False
         self._estimated_num_output_bundles = None
         self._estimated_output_num_rows = None
         self._execution_finished = False
@@ -349,8 +351,7 @@ class PhysicalOperator(Operator):
         self._logical_operators: List[LogicalOperator] = []
         self._data_context = data_context
         self._id = str(uuid.uuid4())
-        # Initialize metrics after data_context is set
-        self._metrics = OpRuntimeMetrics(self)
+        self._metrics = BaseOpMetrics(data_context)
 
     def __reduce__(self):
         raise ValueError("Operator is not serializable.")
@@ -417,13 +418,19 @@ class PhysicalOperator(Operator):
             * The operator has finished execution (i.e., `execution_finished()` is True).
             * All outputs have been taken (i.e., `has_next()` is False) from it.
         """
-        from ..operators.base_physical_operator import InternalQueueOperatorMixin
+        from ray.data._internal.execution.interfaces.op_runtime_metrics import (
+            QueuedOpMetrics,
+        )
 
         internal_input_queue_num_blocks = 0
         internal_output_queue_num_blocks = 0
-        if isinstance(self, InternalQueueOperatorMixin):
-            internal_input_queue_num_blocks = self.internal_input_queue_num_blocks()
-            internal_output_queue_num_blocks = self.internal_output_queue_num_blocks()
+        if isinstance(self.metrics, QueuedOpMetrics):
+            internal_input_queue_num_blocks = (
+                self.metrics.obj_store_mem_internal_inqueue_blocks
+            )
+            internal_output_queue_num_blocks = (
+                self.metrics.obj_store_mem_internal_outqueue_blocks
+            )
 
         if not self._execution_finished:
             if (
@@ -452,7 +459,7 @@ class PhysicalOperator(Operator):
         raise NotImplementedError
 
     @property
-    def metrics(self) -> OpRuntimeMetrics:
+    def metrics(self) -> BaseOpMetrics:
         """Returns the runtime metrics of this operator."""
         self._metrics._extra_metrics = self._extra_metrics()
         return self._metrics
@@ -727,30 +734,6 @@ class PhysicalOperator(Operator):
         """
         return ExecutionResources()
 
-    def notify_in_task_submission_backpressure(self, in_backpressure: bool) -> None:
-        """Called periodically from the executor to update internal in backpressure
-        status for stats collection purposes.
-
-        Args:
-            in_backpressure: Value this operator's in_backpressure should be set to.
-        """
-        # only update on change to in_backpressure
-        if self._in_task_submission_backpressure != in_backpressure:
-            self._metrics.on_toggle_task_submission_backpressure(in_backpressure)
-            self._in_task_submission_backpressure = in_backpressure
-
-    def notify_in_task_output_backpressure(self, in_backpressure: bool) -> None:
-        """Called periodically from the executor to update internal output backpressure
-        status for stats collection purposes.
-
-        Args:
-            in_backpressure: Value this operator's output backpressure should be set to.
-        """
-        # only update on change to in_backpressure
-        if self._in_task_output_backpressure != in_backpressure:
-            self._metrics.on_toggle_task_output_backpressure(in_backpressure)
-            self._in_task_output_backpressure = in_backpressure
-
     def get_autoscaling_actor_pools(self) -> List[AutoscalingActorPool]:
         """Return a list of `AutoscalingActorPool`s managed by this operator."""
         return []
@@ -829,7 +812,7 @@ class ReportsExtraResourceUsage(abc.ABC):
 def estimate_total_num_of_blocks(
     num_tasks_submitted: int,
     upstream_op_num_outputs: int,
-    metrics: OpRuntimeMetrics,
+    metrics: BaseOpMetrics,
     total_num_tasks: Optional[int] = None,
 ) -> Tuple[int, int, int]:
     """This method is trying to estimate total number of blocks/rows based on
