+++ python/ray/data/_internal/execution/operators/map_operator.py
@@ -34,6 +34,9 @@ from ray.data._internal.execution.interfaces import (
     RefBundle,
     TaskContext,
 )
+from ray.data._internal.execution.interfaces.op_runtime_metrics import (
+    TaskOpMetrics,
+)
 from ray.data._internal.execution.interfaces.physical_operator import (
     DataOpTask,
     MetadataOpTask,
@@ -160,7 +163,8 @@ class MapOperator(InternalQueueOperatorMixin, OneToOneOperator, ABC):
         self._next_metadata_task_idx = 0
         # Keep track of all finished streaming generators.
         super().__init__(name, input_op, data_context, target_max_block_size_override)
-
+        # Initialize metrics directly for proper type inference
+        self._metrics = TaskOpMetrics(data_context)
         # If set, then all output blocks will be split into
         # this many sub-blocks. This is to avoid having
         # too-large blocks, which may reduce parallelism for
@@ -193,18 +197,6 @@ class MapOperator(InternalQueueOperatorMixin, OneToOneOperator, ABC):
 
     def set_additional_split_factor(self, k: int):
         self._additional_split_factor = k
-
-    def internal_input_queue_num_blocks(self) -> int:
-        return self._block_ref_bundler.num_blocks()
-
-    def internal_input_queue_num_bytes(self) -> int:
-        return self._block_ref_bundler.size_bytes()
-
-    def internal_output_queue_num_blocks(self) -> int:
-        return self._output_queue.num_blocks()
-
-    def internal_output_queue_num_bytes(self) -> int:
-        return self._output_queue.size_bytes()
 
     def clear_internal_input_queue(self) -> None:
         """Clear internal input queue (block ref bundler)."""
@@ -498,7 +490,8 @@ class MapOperator(InternalQueueOperatorMixin, OneToOneOperator, ABC):
             self._metrics.on_output_queued(output)
 
         def _task_done_callback(task_index: int, exception: Optional[Exception]):
-            self._metrics.on_task_finished(task_index, exception)
+            actor_info = self.get_actor_info()
+            self._metrics.on_task_finished(task_index, exception, actor_info=actor_info)
 
             # Estimate number of tasks and rows from inputs received and tasks
             # submitted so far
