+++ python/ray/data/tests/test_resource_manager.py
@@ -246,11 +246,21 @@ class TestResourceManager:
                 return_value=ExecutionResources.zero()
             )
             op.extra_resource_usage = MagicMock(return_value=ExecutionResources.zero())
+
+            # Mock get_object_store_usage_details() to return ObjectStoreUsageDetails
+            from ray.data._internal.execution.interfaces.op_runtime_metrics import (
+                ObjectStoreUsageDetails,
+            )
+
             op._metrics = MagicMock(
-                obj_store_mem_pending_task_outputs=mock_pending_task_outputs[op],
-                obj_store_mem_internal_outqueue=mock_internal_outqueue[op],
-                obj_store_mem_internal_inqueue=mock_internal_inqueue[op],
-                obj_store_mem_pending_task_inputs=mock_pending_task_inputs[op],
+                get_object_store_usage_details=MagicMock(
+                    return_value=ObjectStoreUsageDetails(
+                        internal_outqueue_memory=mock_internal_outqueue[op],
+                        internal_inqueue_memory=mock_internal_inqueue[op],
+                        pending_task_outputs_memory=mock_pending_task_outputs[op],
+                        pending_task_inputs_memory=mock_pending_task_inputs[op],
+                    )
+                ),
             )
             ref_bundle = MagicMock(
                 size_bytes=MagicMock(return_value=mock_external_outqueue_sizes[op])
@@ -630,9 +640,10 @@ class TestReservationOpResourceAllocator:
             assert allocator._reserved_for_op_outputs[op] == 50
 
     @patch(
-        "ray.data._internal.execution.interfaces.op_runtime_metrics.OpRuntimeMetrics."
-        "obj_store_mem_max_pending_output_per_task",
-        new_callable=PropertyMock(return_value=100),
+        "ray.data._internal.execution.interfaces.op_runtime_metrics.task."
+        "TaskOpMetrics.obj_store_mem_max_pending_output_per_task",
+        new_callable=PropertyMock,
+        return_value=100,
     )
     def test_reserve_min_resources_for_gpu_ops(
         self, mock_property, restore_data_context
